Resolve pytest conflicts
