+++ tests/database/test_database_courses.py
