+++ tests/routers/test_router_courses.py
